refactor: moved helpders to api folder
